Cleanup function param freshening code
